removed MapProxy
